+++ src/controllers/subscription.controller.js
@@ -2,9 +2,16 @@ import { asyncHandler } from "../utils/asyncHandler.js";
 import { Subscription } from "../models/subscription.model.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { ApiError } from "../utils/ApiError.js";
+import { User } from "../models/user.model.js";
 
 const toggleSubscription = asyncHandler(async (req, res) => {
   const { channelId } = req.params;
+  if (channelId === req.user._id.toString()) {
+    throw new ApiError(400, "You cannot subscribe to your own channel");
+  }
+  if (!(await User.exists({ _id: channelId }))) {
+    throw new ApiError(404, "Channel not found");
+  }
   const alreadySubscribed = await Subscription.findOneAndDelete({
     channel: channelId,
     subscriber: req.user._id,
@@ -34,7 +41,8 @@ const getUserChannelSubscribers = asyncHandler(async (req, res) => {
     throw new ApiError(400, "No channelId is Provided");
   }
   const subscribers = await Subscription.find({ channel: channelId }).populate(
-    "subscriber"
+    "subscriber",
+    "-password -refreshToken"
   );
   return res
     .status(200)
@@ -48,7 +56,7 @@ const getSubscribedChannels = asyncHandler(async (req, res) => {
   }
   const channels = await Subscription.find({
     subscriber: subscriberId,
-  }).populate("channel");
+  }).populate("channel", "-password -refreshToken");
   return res
     .status(200)
     .json(
